pushing tough bricks
I just changed what the bone colors ment. Will work on a more robust way of saying what we want to stay the same. Probably a second layer that is overlayed on top of the boss that looks one layer under and tells the brick its toughness. that way toughness looks like a heat map